backend/requirements.txt
beautifulsoup4
firebase-admin
pandas
python-dotenv
requests
